+++ wxcadm/__init__.py
@@ -39,3 +39,4 @@ from .reports import *
 from .virtual_line import *
 from .webhooks import *
 from .workspace import *
+from .pstn import *